fix: column-edit-menu component split
