+++ src/components/index.ts
@@ -1,6 +1,5 @@
 export * from './TableInformation';
 export * from './RelationshipInformation';
 export * from './ErdDrawTools';
-export * from './ColumnEditMenu';
 export * from './Relationship';
 export * from './ConnectLine';
